version 1.3 scribble works

+++ gazeplay/src/main/java/net/gazeplay/commons/utils/Position.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class Position {
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 }
